refactor: auth service to register service
