+++ src/services/register-service.ts
@@ -1,5 +1,4 @@
 import {
-  createAuthForClientCredentialsFlow,
   createHttpClient,
   createClient,
   createAuthForAnonymousSessionFlow,
@@ -12,52 +11,7 @@ import { type ApiError } from '../types/api-error';
 import { type Address } from '../types/address';
 
 const projectKey = 'rss-ecom';
-//const authStore = useAuthStore();
 
-const client = createClient({
-  middlewares: [
-    createAuthForClientCredentialsFlow({
-      host: 'https://auth.us-central1.gcp.commercetools.com',
-      projectKey,
-      credentials: {
-        clientId: import.meta.env.VITE_CTP_CLIENT_ID,
-        clientSecret: import.meta.env.VITE_CTP_CLIENT_SECRET,
-      },
-    }),
-    createHttpClient({
-      host: 'https://api.us-central1.gcp.commercetools.com',
-      fetch: globalThis.fetch,
-    }),
-  ],
-});
-
-export const apiRoot = createApiBuilderFromCtpClient(client).withProjectKey({ projectKey });
-
-export const loginCustomer = async (
-  email: string,
-  password: string,
-  authStore: ReturnType<typeof useAuthStore>,
-): Promise<void> => {
-  try {
-    await apiRoot
-      .login()
-      .post({
-        body: {
-          email,
-          password,
-        },
-      })
-      .execute();
-
-    authStore.setUser(email, password);
-    authStore.setAuth(true);
-  } catch (error: unknown) {
-    const defaultError = 'Server authentication error';
-    const errorMessage = isCorrectError(error) ? error.message : defaultError;
-
-    authStore.setError(errorMessage);
-  }
-};
 function isCorrectError(error: unknown): error is ApiError {
   if (typeof error !== 'object' || error === null) return false;
